pipeline/requirements.txt
requests
pandas
psycopg2-binary
python-dotenv
sqlalchemy
dbt-postgres